+++ package.json
@@ -39,7 +39,7 @@
     }
   },
   "scripts": {
-    "test": "mocha",
+    "test": "mocha --require babel-register tests/*.js --exit",
     "lint": "eslint",
     "start": "DEBUG=app nodemon app.js --exec babel-node --"
   },
